Revert "SomeChanges"
This reverts commit 9097a7b76fe54518e105b554a798f2d0577ff5c8.

+++ RestaurantTemplate/Restaurant.Core/Order.cs
@@ -8,23 +8,24 @@ namespace Restaurant.Core
 {
     public class Order
     {
-        public Order(double delay, string guestName, OrderType orderType)
+        public Order(int delay, string guestName, OrderType orderType)
         {
             Delay = delay;
             GuestName = guestName;
             OrderType = orderType;
         }
-        public Order(double delay, string guestName, OrderType orderType, Article article) :this (delay,guestName, orderType)
+        public Order(int delay, string guestName, OrderType orderType, Article article) :this (delay,guestName, orderType)
         {
             Article = article;
         }
 
-        public double Delay { get; set; }
+        public int Delay { get; set; }
 
         public string GuestName { get; set; }
 
         public OrderType OrderType { get; set; }
        
+
         public Article Article{ get; set; }
     }
 }
